# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.EditRatingRequest:
    properties:
      stars:
        type: integer
      user-id:
        type: string
    required:
    - stars
    - user-id
    type: object
...
  dto.RecordDiscordRequest:
    properties:
      discord:
        type: string
      user-id:
        type: string
    required:
    - discord
    - user-id
    type: object
...
  dto.UploadAvatarRequest:
    type: object
...
  entities.User:
    properties:
      avatar:
        type: string
      chat_id:
        type: string
      discord:
        type: string
      games:
        items:
          type: string
        type: array
      id:
        type: string
      login:
        type: string
      num_of_ratings:
        type: integer
      rating:
        type: number
      telegram:
        type: string
      total_rating:
        type: integer
    type: object
host: localhost:1111
info:
  contact: {}
  description: Ненавижу эту хуйню бля
  title: На жри уебок доки свои не подавись (недоделанная кстати хихиххихихих))))))
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users
      parameters:
      - in: query
        name: amount
        required: true
        type: integer
      - in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get users list
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Upload or update user avatar image
      parameters:
      - description: Avatar upload data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UploadAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upload user avatar
      tags:
      - users
  /users/avatar/{id}:
    delete:
      consumes:
      - application/json
      description: Remove user's avatar image
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user avatar
      tags:
      - users
  /users/discord:
    post:
      consumes:
      - application/json
      description: Link user account with Discord
      parameters:
      - description: Discord association data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordDiscordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Record Discord association
      tags:
      - users
  /users/rating:
    patch:
      consumes:
      - application/json
      description: Update user's rating value
      parameters:
      - description: Rating edit data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Edit user rating
      tags:
      - users
swagger: "2.0"